Refactor auth pages and context imports
