+++ src/contexts/auth/index.js
@@ -1,0 +1,1 @@
+export { AuthProvider, useAuth } from './AuthContext';
