+++ src/pages/OAuthCallback.jsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import { useNavigate, useSearchParams } from 'react-router-dom';
-import { useAuth } from '../contexts/AuthContext';
+import { useAuth } from '@/contexts';
 import { Bot } from 'lucide-react';
 
 export const OAuthCallback = () => {
